scripts/requirements.txt
psycopg2-binary==2.9.7
faker==19.6.2
python-dotenv==1.0.0
tqdm==4.66.1
